Move port::test module into separate file
No tests were changed, just moved.
